+++ crucible-concurrency/src/Cruces/ExploreCrux.hs
@@ -5,14 +5,23 @@ Copyright        : (c) Galois, Inc 2021
 Maintainer       : Alexander Bakst <abakst@galois.com>
 -}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE ImplicitParams #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators #-}
 
-module Cruces.ExploreCrux where
+module Cruces.ExploreCrux
+  ( CruxPersonality
+  , mkCruxPersonality
+  , exploreCallback
+  , exploreOvr
+  ) where
 
 import           Control.Monad.IO.Class
 import           Control.Monad (when)
@@ -20,12 +29,15 @@ import           Control.Lens
 import           Data.Generics.Product.Fields (field, setField)
 import qualified Data.Vector as V
 import qualified Data.Map.Strict as Map
+import           Data.Void (Void)
 import           System.IO (Handle)
 
 import           What4.Interface
 import           What4.Config
 
 import qualified Data.Parameterized.Context as Ctx
+import qualified Data.Parameterized.Map as MapF
+import qualified Lang.Crucible.Debug as Debug
 import           Lang.Crucible.FunctionHandle (HandleAllocator)
 import qualified Lang.Crucible.CFG.Core as C
 import           Lang.Crucible.Simulator
@@ -46,6 +58,60 @@ import           Crucibles.Common
 import           Crux.Goal (proveGoalsOnline)
 import           Crux.Types (totalProcessedGoals, provedGoals)
 
+
+-- | A Crucible personality type
+-- (see 'Lang.Crucible.Simulator.ExecutionTree.cruciblePersonality')
+-- suitable for use with @crucible-concurrency@ and Crux.
+data CruxPersonality alg ext ret sym
+  = CruxPersonality
+    { _pExploration :: Exploration (CruxPersonality alg ext ret sym) alg ext ret sym
+    , _pDbgContext :: Debug.Context Void sym ext ret
+    }
+makeLenses ''CruxPersonality
+
+instance HasExploration (CruxPersonality alg ext ret sym) alg ext ret sym where
+  exploration = pExploration
+  {-# INLINE exploration #-}
+
+instance Debug.HasContext (CruxPersonality alg ext ret sym) Void sym ext ret where
+  context = pDbgContext
+  {-# INLINE context #-}
+
+mkCruxPersonality ::
+  SchedulingAlgorithm alg =>
+  IsExpr (SymExpr sym) =>
+  IO (CruxPersonality alg ext C.UnitType sym)
+mkCruxPersonality = do
+  let cExts = Debug.voidExts
+  inps <- Debug.defaultDebuggerInputs cExts
+  dbgCtx <-
+    Debug.initCtx
+      cExts
+      (Debug.IntrinsicPrinters MapF.empty)
+      inps
+      Debug.defaultDebuggerOutputs
+      C.UnitRepr
+  pure (CruxPersonality emptyExploration dbgCtx)
+  where
+    emptyExploration ::
+      SchedulingAlgorithm alg =>
+      Exploration p alg ext C.UnitType sym
+    emptyExploration =
+      Exploration
+      { _exec      = initialExecutions
+      , _scheduler = s0
+      , _schedAlg  = initialAlgState
+      , _num       = 0
+      , _gVars     = mempty
+      }
+      where
+        s0 = Scheduler { _threads      = V.fromList [EmptyThread]
+                       , _retRepr      = C.UnitRepr
+                       , mainCont      = return ()
+                       , _activeThread = 0
+                       , _numSwitches  = 0
+                       }
+
 -- | Callback for crucible-syntax exploration
 exploreCallback :: forall alg st.
   (?bound::Int, SchedulingAlgorithm alg) =>
@@ -57,9 +123,9 @@ exploreCallback :: forall alg st.
    IsSymBackend s bak =>
    bak ->
         IO ( FnVal s Ctx.EmptyCtx C.UnitType
-           , ExplorePrimitives (Personality alg () C.UnitType s) s ()
+           , ExplorePrimitives (CruxPersonality alg () C.UnitType s) s ()
            , [Some C.GlobalVar]
-           , FunctionBindings (Personality alg () C.UnitType s) s ())
+           , FunctionBindings (CruxPersonality alg () C.UnitType s) s ())
            ) ->
   Crux.SimulatorCallbacks Crux.CruxLogMessage st Crux.Types.CruxSimulationResult
 exploreCallback cruxOpts ha outh mkSym =
@@ -70,8 +136,8 @@ exploreCallback cruxOpts ha outh mkSym =
         { Crux.setupHook =
             \bak symOnline ->
               do (mainHdl, prims, globs, fns) <- mkSym bak
-                 let simCtx = initSimContext bak emptyIntrinsicTypes ha outh fns emptyExtensionImpl (Personality emptyExploration)
-
+                 p <- mkCruxPersonality 
+                 let simCtx = initSimContext bak emptyIntrinsicTypes ha outh fns emptyExtensionImpl p
                      st0  = InitialState simCtx emptyGlobals defaultAbortHandler C.UnitRepr $
                                 runOverrideSim
                                   C.UnitRepr
@@ -87,23 +153,6 @@ exploreCallback cruxOpts ha outh mkSym =
         , Crux.onErrorHook = \_bak -> return (\_ _ -> return mempty)
         , Crux.resultHook = \_bak result -> return result
         }
-
-
--- | Empty exploration state
-emptyExploration :: SchedulingAlgorithm alg => Exploration p alg ext C.UnitType sym
-emptyExploration = Exploration { _exec      = initialExecutions
-                               , _scheduler = s0
-                               , _schedAlg  = initialAlgState
-                               , _num       = 0
-                               , _gVars     = mempty
-                               }
-  where
-    s0 = Scheduler { _threads      = V.fromList [EmptyThread]
-                   , _retRepr      = C.UnitRepr
-                   , mainCont      = return ()
-                   , _activeThread = 0
-                   , _numSwitches  = 0
-                   }
 
 -- | Wrap an override to produce a NEW override that will explore the executions of a concurrent program.
 -- Must also use the 'scheduleFeature' 'ExecutionFeature'