+++ crux/src/Crux.hs
@@ -57,7 +57,6 @@ import           System.FilePath ((</>))
 import           System.IO ( Handle, hPutStr, stdout, stderr )
 
 import           Data.Parameterized.Classes
-import qualified Data.Parameterized.Map as MapF
 import           Data.Parameterized.Nonce (newIONonceGenerator, NonceGenerator)
 import           Data.Parameterized.Some ( Some(..) )
 
@@ -102,13 +101,20 @@ import           Crux.Log as Log
 import           Crux.Report
 import           Crux.Types
 
-pattern RunnableState :: forall sym . () => forall ext personality . (IsSyntaxExtension ext) => ExecState (personality sym) sym ext (RegEntry sym UnitType) -> RunnableState sym
+pattern RunnableState ::
+  forall sym. () =>
+  forall ext personality.
+  ( IsSyntaxExtension ext
+  , Debug.HasContext (personality sym) Void sym ext UnitType
+  ) =>
+  ExecState (personality sym) sym ext (RegEntry sym UnitType) ->
+  RunnableState sym
 pattern RunnableState es = RunnableStateWithExtensions es []
 
 -- | A crucible @ExecState@ that is ready to be passed into the simulator.
 --   This will usually, but not necessarily, be an @InitialState@.
 data RunnableState sym where
-  RunnableStateWithExtensions :: (IsSyntaxExtension ext)
+  RunnableStateWithExtensions :: (IsSyntaxExtension ext, Debug.HasContext (personality sym) Void sym ext UnitType)
                               => ExecState (personality sym) sym ext (RegEntry sym UnitType)
                               -> [ExecutionFeature (personality sym) sym ext (RegEntry sym UnitType)]
                               -> RunnableState sym
@@ -721,17 +727,7 @@ doSimWithResults cruxOpts simCallback bak execFeatures profInfo monline goalProv
       if debugging
       then do
         let ?parserHooks = Syn.ParserHooks Applicative.empty Applicative.empty
-        let cExts = Debug.voidExts
-        inps <- Debug.defaultDebuggerInputs cExts
-        dbg <-
-          Debug.debugger
-            cExts
-            Debug.voidImpl
-            (Debug.IntrinsicPrinters MapF.empty)
-            inps
-            Debug.defaultDebuggerOutputs
-            UnitRepr
-        pure [dbg]
+        pure [Debug.debugger Debug.voidImpl]
       else pure []
 
     -- execute the simulator