+++ crucible-debug/CHANGELOG.md
@@ -1,6 +1,6 @@
 # next
 
-Nothing yet.
+- Store debugger state in the Crucible personality.
 
 # 0.1 -- 2025-03-21